requirements.txt
numpy
pandas
librosa
scikit-learn
joblib
matplotlib
seaborn
pyyaml
fastapi
uvicorn
python-multipart